Part 1 initial attempt
Decided against this way because it was going to be a nightmare to implement. I was trying to find all the symbols, then check all the adjacent/diagonal indeces and then if one of them was a digit walk through the adjacent indeces to form a number and log it.

Decided at this point to just throw this out and do the opposite instead and compile numbers together and see if there is a symbol in an adjacent or diagonal index.